Use "legacy" filenames / change ScopeID setting
The filenames for the SI and PI files have been changed to match the formatting of the pre-specification format. This should not affect the operation on devices adhering the to the published standards.

The ScopeID of the PI files has been extended to add SId and SCIds values, which were previously omitted.